Add redirects for Netlify
We added a 'netlify.toml' file (a configuration file) to the root of the
project for redirecting '/' to '/home', which was showing an error page
earlier. We could have also added a '_redirects' file instead, but
having the netlify.toml file simplifies maintainence since all settings
can exist in the same file.

We used the redirect which can be used with most SPAs (Single Page
Application) that handle routing internally.

Close #27